+++ Core/Entities/Inventario.cs
@@ -9,9 +9,9 @@ public class Inventario : BaseEntityStr
 {
     public string Nombre { get; set; }
     public double Precio { get; set; }
-    public short Stock { get; set; }
-    public short StockMin { get; set; }
-    public short StockMax { get; set; }
+    public int Stock { get; set; }
+    public int StockMin { get; set; }
+    public int StockMax { get; set; }
     public int IdPresentacion { get; set; }
     public Presentacion Presentaciones { get; set; }
     public string CodProducto { get; set; }
